requirements.txt
dataclasses_json>=0.5.7
python-dotenv>=1.0.0
requests>=2.28.1
tqdm>=4.64.1
configparser>=5.3.0
